bug 20791: fixed a problem with getting negative distances with Levenshtein + fixed rounding issues with (double-)metaphone algorithms
git-svn-id: http://talendforge.org/svn/top/trunk@59274 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.dataquality.record.linkage/src/main/java/org/talend/dataquality/record/linkage/attribute/DoubleMetaphoneMatcher.java
@@ -36,7 +36,8 @@ public class DoubleMetaphoneMatcher implements IAttributeMatcher {
      * @see org.talend.dataquality.record.linkage.attribute.IAttributeMatcher#getMatchingWeight(java.lang.String, java.lang.String)
      */
     public double getMatchingWeight(String str1, String str2) {
-        return StringComparisonUtil.difference(algorithm.encode(str1), algorithm.encode(str2)) / algorithm.getMaxCodeLen();
+        return StringComparisonUtil.difference(algorithm.encode(str1), algorithm.encode(str2))
+                / (double) algorithm.getMaxCodeLen();
     }
 
 }
